+++ docs/CHANGELOG.md
@@ -1,3 +1,15 @@
+Version 3.36.0 (July 5, 2023)
+-----------------------------
+
+**User-visible changes:**
+
+**Implementation details:**
+
+Deprecated `ElementUtils.getSimpleNameOrDescription()` in favor of `getSimpleDescription()`.
+
+**Closed issues:**
+
+
 Version 3.35.0 (June 1, 2023)
 ------------------------------